Removed argument injection via method
