+++ MANUAL.md
@@ -18,31 +18,27 @@ Test classes support both `Argument` injection (`@TestEngine.Argument` annotated
 - `@TestEngine.AfterEach`
 - `@TestEngine.AfterAll`
 
-`@TestEngine.Argument` field injection example:
+`@TestEngine.Argument` member argument example:
 
 - [SimpleTest.java](/examples/src/test/java/example/SimpleTest.java)
 
-`@TestEngine.Argument` method injection example:
+`@TestEngine` test method parameter example:
 
-- [SimpleTest2.java](/examples/src/test/java/example/SimpleTest2.java)
-
-`@TestEngine` test method argument example:
-
-- [SimpleTest3.java](/examples/src/test/java/example/SimpleTest3.java)
+- [NoArgumentDeclarationTest.java](/examples/src/test/java/example/NoArgumentDeclarationTest.java)
 
 ### Test Annotations
 
-| Annotation                     | Static | Type            | Required | Example                                                                                                                                                                            |
-|--------------------------------|--------|-----------------|----------|------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|
-| `@TestEngine.ArgumentSupplier` | yes    | method          | yes      | <nobr>`public static Stream<[Object that implements Argument]> arguments();`</nobr><br/><br/><nobr>`public static Iterable<[Object that implements Argument]> arguments();`</nobr> |
-| `@TestEngine.Argument`         | no     | field or method | no       | Note: can be used on a field or method                                                                                                                                             |
-| `@TestEngine.Prepare`          | no     | method          | no       | `public void prepare();`                                                                                                                                                           |
-| `@TestEngine.BeforeAll`        | no     | method          | no       | `public void beforeAll();`                                                                                                                                                         |
-| `@TestEngine.BeforeEach`       | no     | method          | no       | `public void beforeEach();`                                                                                                                                                        |
-| `@TestEngine.Test`             | no     | method          | yes      | `public void test();` or `public void test(Argument argument)`                                                                                                                     |
-| `@TestEngine.AfterEach`        | no     | method          | no       | `public void afterEach();`                                                                                                                                                         |
-| `@TestEngine.AfterAll`         | no     | method          | no       | `public void afterAll();`                                                                                                                                                          |
-| `@TestEngine.Conclude`         | no     | method          | no       | `public void conclude();`                                                                                                                                                          |
+| Annotation                     | Static | Type   | Required | Example                                                                                                                                                                            |
+|--------------------------------|--------|--------|----------|------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|
+| `@TestEngine.ArgumentSupplier` | yes    | method | yes      | <nobr>`public static Stream<[Object that implements Argument]> arguments();`</nobr><br/><br/><nobr>`public static Iterable<[Object that implements Argument]> arguments();`</nobr> |
+| `@TestEngine.Argument`         | no     | field  | no       | `public Argument argument;`                                                                                                                                                        |
+| `@TestEngine.Prepare`          | no     | method | no       | `public void prepare();`                                                                                                                                                           |
+| `@TestEngine.BeforeAll`        | no     | method | no       | `public void beforeAll();`                                                                                                                                                         |
+| `@TestEngine.BeforeEach`       | no     | method | no       | `public void beforeEach();`                                                                                                                                                        |
+| `@TestEngine.Test`             | no     | method | yes      | `public void test();` or `public void test(Argument argument)`                                                                                                                     |
+| `@TestEngine.AfterEach`        | no     | method | no       | `public void afterEach();`                                                                                                                                                         |
+| `@TestEngine.AfterAll`         | no     | method | no       | `public void afterAll();`                                                                                                                                                          |
+| `@TestEngine.Conclude`         | no     | method | no       | `public void conclude();`                                                                                                                                                          |
 
 Reference the [Design](https://github.com/antublue/test-engine#design) for the test engine execution flow.
